+++ tests/key.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Test\Key;
+use StarkInfra\Key;
+
+
+class TestKey
+{
+  public function create()
+  {
+    list($priv, $pub) = Key::create();
+    list($priv2, $pub2) = Key::create("keys");
+  }
+}
+
+echo "\n\nKey:";
+
+$test = new TestKey();
+
+echo "\n\t- create";
+$test->create();
+echo " - OK";